Use proper namespace for elliptic interface  AL-based preconditioner
